Separate Modal CSV component into jsx file
